requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
pydantic==2.11.7
python-dotenv==1.1.1
email-validator==2.2.0
pytest==8.4.1
